requirements.txt
altair==5.1.2
anyio==3.7.1
appnope==0.1.3
arch==7.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.2.0
clang==17.0.6
cleo==2.1.0
click==8.1.7
comm==0.1.3
contourpy==1.2.0
crashtest==0.4.1
cycler==0.12.1
Cython==3.0.10
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
dulwich==0.21.7
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.13.4
fonttools==4.45.1
fqdn==1.5.1
frozendict==2.4.5
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
ipykernel==6.25.0
ipython==8.27.0
ipywidgets==8.1.3
isoduration==20.11.0
jaraco.classes==3.4.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.2.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
keyring==24.3.1
kiwisolver==1.4.5
lib==4.0.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mlkit==0.0.1
more-itertools==10.2.0
msgpack==1.0.8
multitasking==0.0.11
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.1
nest-asyncio==1.5.6
notebook==7.2.1
notebook_shim==0.2.3
numpy==1.25.2
overrides==7.3.1
packaging==23.1
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
peewee==3.17.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pkginfo==1.10.0
platformdirs==3.9.1
plotly==5.24.1
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
prometheus-client==0.17.1
prompt_toolkit==3.0.48
protobuf==4.25.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pycparser==2.21
pydeck==0.8.1b0
pyemd==1.0.0
Pygments==2.15.1
pykalman==0.9.7
pyparsing==3.1.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-telegram-bot==21.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.5.2
QtPy==2.4.1
rapidfuzz==3.8.1
referencing==0.30.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.9.2
ruptures==1.1.9
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
sklearn==0.0.post11
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.28.2
ta==0.11.0
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
tomlkit==0.12.4
toolz==0.12.0
tornado==6.3.2
traitlets==5.14.3
trove-classifiers==2024.4.10
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
urllib3==2.0.4
validators==0.22.0
virtualenv==20.26.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.11
xattr==1.1.0
yfinance==0.2.44
zipp==3.17.0
